web_ui/requirements.txt
apipkg==1.5
Appium-Python-Client==0.52
atomicwrites==1.2.1
attrs==18.2.0
bdd==0.2
behave==1.2.5
castro==1.0.6
certifi==2018.10.15
cffi==1.14.4
chardet==3.0.4
colorama==0.3.9
cryptography==3.3.1
ddt==1.2.0
enum34==1.1.6
execnet==1.5.0
glob2==0.7
httpagentparser==1.8.1
idna==2.7
Logbook==1.5.3
lxml==4.6.2
mailosaur==5.0.5
Mako==1.1.0
MarkupSafe==1.1.1
more-itertools==4.3.0
namedlist==1.7
names==0.3.0
nose==1.3.7
parse==1.9.0
parse-type==0.4.2
pathlib2==2.3.2
pep8==1.7.1
pluggy==0.7.1
py==1.6.0
pycparser==2.20
pygame==1.9.6
pyOpenSSL==20.0.1
pytest==3.8.1
pytest-allure-adaptor==1.7.10
pytest-bdd==3.2.1
pytest-cache==1.0
pytest-html==1.19.0
pytest-logbook==1.2.0
pytest-metadata==1.7.0
pytest-pep8==1.0.6
pytest-repeat==0.7.0
python-dateutil==2.7.3
PyYAML==5.2
requests==2.20.0
scandir==1.9.0
selenium==3.14.1
six==1.11.0
typing==3.7.4.1
urllib3==1.23
virtualenv==16.0.0
